fix: add __attribute__((format(printf))) to kprintf family, fix all format specifiers in modified files
- Add __attribute__((format(printf))) to kprintf, vkprintf, snprintf, vsnprintf, sprintf
- Fix format specifiers throughout: %x -> %llx for 64-bit values, remove unnecessary
  (uint64_t) casts for small integer types
- Fix kernel.c: remove (uint64_t) casts on uint8_t/uint32_t fields (rtc, mac, ip)
- Fix fault.c: use %llx for all 64-bit register values
- Fix ata.c: use %llu for sector counts
- Add heap.h to check-app-boundary allowlist (for cmd_dmesg.c kmalloc usage)
- cmd_dmesg.c: use heap-allocated buffer instead of 16KB stack buffer

+++ src/drivers/ata.c
@@ -106,8 +106,8 @@ void ata_init(void) {
     uint32_t sectors = identify[60] | ((uint32_t)identify[61] << 16);
     ata_total_sectors = sectors;
     blockdev_register(BLOCKDEV_ATA, "ata", ata_read_sectors, ata_write_sectors, ata_get_sectors);
-    kprintf("  ATA disk: %u sectors (%u MB)\n", (uint64_t)sectors,
-            (uint64_t)(sectors / 2048));
+    kprintf("  ATA disk: %llu sectors (%llu MB)\n", (unsigned long long)sectors,
+            (unsigned long long)(sectors / 2048));
 }
 
 int ata_is_present(void) {
